move location of tests and update imports accordingly
